test: update to parameterized test suite
